iio: adc: at91_adc: allow to use full range of startup time
commit 2ab5f39bc7825808e0fa1e7e5f0b23e174563467 upstream.

The DT-Property "atmel,adc-startup-time" is stored in an u8 for a microsecond
value. When trying to increase the value of STARTUP in Register AT91_ADC_MR
some higher values can't be reached.

Change the type in function parameter and private structure field from u8 to
u32.

[nicolas.ferre@atmel.com: change commit message, increase u16 to u32 for startup time]
Acked-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>

+++ drivers/iio/adc/at91_adc.c
@@ -55,7 +55,7 @@ struct at91_adc_state {
 	u8			num_channels;
 	void __iomem		*reg_base;
 	struct at91_adc_reg_desc *registers;
-	u8			startup_time;
+	u32			startup_time;
 	u8			sample_hold_time;
 	bool			sleep_mode;
 	struct iio_trigger	**trig;
